requirements.txt
google-cloud-logging==3.9.0
google-cloud-storage==2.14.0
gradio==4.12.0
Pillow==10.1.0
google-cloud-aiplatform==1.44.0
moviepy==1.0.3
db-dtypes==1.2.0